♻️  Refactor : 에러 처리 service단으로 이동

+++ src/main/java/backend/like_house/domain/user_management/service/impl/UserManagementCommandServiceImpl.java
@@ -1,6 +1,7 @@
 package backend.like_house.domain.user_management.service.impl;
 
 import backend.like_house.domain.user.entity.User;
+import backend.like_house.domain.user.repository.UserRepository;
 import backend.like_house.domain.user_management.converter.UserManagementConverter;
 import backend.like_house.domain.user_management.dto.UserManagementDTO.UserManagementRequest.ModifyFamilyDataRequest;
 import backend.like_house.domain.user_management.entity.Contact;
@@ -9,6 +10,10 @@ import backend.like_house.domain.user_management.repository.BlockUserRepository;
 import backend.like_house.domain.user_management.repository.ContactRepository;
 import backend.like_house.domain.user_management.repository.CustomRepository;
 import backend.like_house.domain.user_management.service.UserManagementCommandService;
+import backend.like_house.global.error.code.status.ErrorStatus;
+import backend.like_house.global.error.handler.FamilySpaceException;
+import backend.like_house.global.error.handler.UserException;
+import backend.like_house.global.error.handler.UserManagementException;
 import java.util.Optional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -22,9 +27,16 @@ public class UserManagementCommandServiceImpl implements UserManagementCommandSe
     private final ContactRepository contactRepository;
     private final CustomRepository customRepository;
     private final BlockUserRepository blockUserRepository;
+    private final UserRepository userRepository;
 
     @Override
     public Custom modifyFamilyCustom(User user, Long userId, ModifyFamilyDataRequest request) {
+        User modifyUser = userRepository.findById(userId)
+                .orElseThrow(() -> new UserException(ErrorStatus.USER_NOT_FOUND));
+        if (!modifyUser.getFamilySpace().equals(user.getFamilySpace())) {
+            throw new FamilySpaceException(ErrorStatus.NOT_INCLUDE_USER_FAMILY_SPACE);
+        }
+
         Optional<Contact> contact = contactRepository.findByUserAndFamilySpaceAndProfileId(
                 user, user.getFamilySpace(), userId);
         Custom custom;
@@ -39,13 +51,33 @@ public class UserManagementCommandServiceImpl implements UserManagementCommandSe
     }
 
     @Override
-    public void blockUser(User manager, User blockUser) {
+    public void blockUser(User manager, Long blockUserId) {
+        User blockUser = userRepository.findById(blockUserId)
+                .orElseThrow(() -> new UserException(ErrorStatus.USER_NOT_FOUND));
+
+        if (blockUserRepository.existsByUserAndFamilySpace(blockUser, manager.getFamilySpace())) {
+            throw new UserManagementException(ErrorStatus.ALREADY_BLOCKED_USER);
+        }
+        if (!blockUser.getFamilySpace().equals(manager.getFamilySpace())) {
+            throw new FamilySpaceException(ErrorStatus.NOT_INCLUDE_USER_FAMILY_SPACE);
+        }
+
         blockUser.setFamilySpace(null);
         blockUserRepository.save(UserManagementConverter.toBlockUser(blockUser, manager.getFamilySpace()));
     }
 
     @Override
-    public void releaseBlockUser(User manager, User blockUser) {
+    public void releaseBlockUser(User manager, Long blockUserId) {
+        User blockUser = userRepository.findById(blockUserId)
+                .orElseThrow(() -> new UserException(ErrorStatus.USER_NOT_FOUND));
+
+        if (!blockUserRepository.existsByUserAndFamilySpace(blockUser, manager.getFamilySpace())) {
+            throw new UserManagementException(ErrorStatus.ALREADY_RELEASE_BLOCK_USER);
+        }
+        if (blockUser.getFamilySpace() != null) {
+            throw new FamilySpaceException(ErrorStatus.ALREADY_BELONG_OTHER_FAMILY_SPACE);
+        }
+
         blockUser.setFamilySpace(manager.getFamilySpace());
         blockUserRepository.deleteByUserAndFamilySpace(blockUser, manager.getFamilySpace());
     }
